SOCINT-266 Use /survey endpoint for available fulfilments Started split out of validators
