+++ app/utils.py
@@ -4,7 +4,6 @@ import re
 from aiohttp.client_exceptions import (ClientResponseError)
 from .exceptions import InvalidDataError, InvalidDataErrorWelsh, TooManyRequestsEQLaunch
 from aiohttp.web import HTTPFound
-from datetime import datetime
 from pytz import timezone
 from unicodedata import normalize
 
@@ -284,52 +283,6 @@ class ProcessName:
             name_valid = False
 
         return name_valid
-
-
-class ProcessDOB:
-    @staticmethod
-    def validate_dob(data):
-        form_day = data.get('day')
-        form_month = data.get('month')
-        form_year = data.get('year')
-
-        try:
-            date = datetime(int(form_year), int(form_month), int(form_day)).date()
-            return date
-        except ValueError:
-            raise InvalidDataError('invalid dob', message_type='invalid')
-
-    @staticmethod
-    def format_dob(date_value):
-        unformatted_date = datetime.strptime(date_value, '%Y-%m-%d')
-        formatted_date = unformatted_date.strftime('%d %B %Y')
-        return formatted_date
-
-
-class ProcessEmailAddress:
-    email_validation_pattern = re.compile(
-        r'(^[^@\s]+@[^@\s]+\.[^@\s]+$)'
-    )
-
-    @staticmethod
-    def validate_email(email, locale):
-        if len(email.strip()) == 0:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Enter an email address", message_type='empty')
-            else:
-                raise InvalidDataError('Enter an email address', message_type='empty')
-        else:
-            if ProcessEmailAddress.email_validation_pattern.fullmatch(email):
-                return email
-            else:
-                if locale == 'cy':
-                    raise InvalidDataErrorWelsh(
-                        "Enter an email address in a valid format, for example, name@example.com",
-                        message_type='invalid')
-                else:
-                    raise InvalidDataError(
-                        'Enter an email address in a valid format, for example, name@example.com',
-                        message_type='invalid')
 
 
 class FlashMessage:
